commented out vendhq

+++ backend/src/main/java/ecommerce/app/backend/markets/squareup/items/SquarePriceMoney.java
@@ -5,7 +5,7 @@ import lombok.Setter;
 
 public class SquarePriceMoney {
     @Getter @Setter
-    private Float amount;
+    private Long amount;
 
     @Getter @Setter
     private String currency;
